Created the beginning of classes stub
Created most of the files we will need and started to fill out some of
them. Added some variables and comments to some of the classes.

Not much but at least it's a start.

Also included some ui ideas I though of. Some are a bit wacky so I
really don't expect/wan't to build them but they add to the list of
ideas we need to show scott